buf: Support big buffers
This commit moves codegen for buf() over to string scratch buffer
infrastructure. Note that we also increase the length field size from 1B
to 4B so that users can take properly advantage of the big buffer
support.

+++ src/ast/async_event_types.h
@@ -68,14 +68,14 @@ struct Strftime {
 } __attribute__((packed));
 
 struct Buf {
-  uint8_t length;
+  uint32_t length;
   // Seems like GCC 7.4.x can't handle `char content[]`. Work around by using
   // 0 sized array (a GCC extension that clang also accepts:
   // https://gcc.gnu.org/bugzilla/show_bug.cgi?id=70932). It also looks like
   // the issue doesn't exist in GCC 7.5.x.
   char content[0];
 
-  std::vector<llvm::Type*> asLLVMType(ast::IRBuilderBPF& b, size_t length);
+  std::vector<llvm::Type*> asLLVMType(ast::IRBuilderBPF& b, uint32_t length);
 } __attribute__((packed));
 
 struct HelperError {
